all: removed omnetpp::cPar::longValue() calls.
Motivation:
rely on cast operators instead because they can detect overflows (when the value is cast to a smaller integer type that can not represent it).
Removed unnecessary .longvalue() calls.
Replaced to .intValue() in comparisons.

+++ src/inet/applications/dhcp/DhcpServer.cc
@@ -523,7 +523,7 @@ void DhcpServer::startApp()
     const char *gatewayStr = par("gateway").stringValue();
     gateway = *gatewayStr ? L3AddressResolver().resolve(gatewayStr, L3AddressResolver::ADDR_IPv4).toIPv4() : ipv4data->getIPAddress();
     subnetMask = ipv4data->getNetmask();
-    long numReservedAddresses = par("numReservedAddresses").longValue();
+    long numReservedAddresses = par("numReservedAddresses");
     uint32_t networkStartAddress = ipv4data->getIPAddress().getInt() & ipv4data->getNetmask().getInt();
     ipAddressStart = Ipv4Address(networkStartAddress + numReservedAddresses);
     if (!Ipv4Address::maskedAddrAreEqual(ipv4data->getIPAddress(), ipAddressStart, subnetMask))
