+++ src/inet/linklayer/csma/Csma.cc
@@ -196,7 +196,7 @@ InterfaceEntry *Csma::createInterfaceEntry()
     e->setInterfaceToken(address.formInterfaceIdentifier());
 
     // capabilities
-    e->setMtu(par("mtu").longValue());
+    e->setMtu(par("mtu"));
     e->setMulticast(true);
     e->setBroadcast(true);
