+++ src/inet/linklayer/ethernet/EtherMacBase.cc
@@ -202,7 +202,7 @@ void EtherMacBase::initializeQueueModule()
             txQueue.extQueue->requestPacket();
     }
     else {
-        txQueue.setInternalQueue("txQueue", par("txQueueLimit").longValue());
+        txQueue.setInternalQueue("txQueue", par("txQueueLimit"));
     }
 }
 
@@ -283,7 +283,7 @@ InterfaceEntry *EtherMacBase::createInterfaceEntry()
 
     // MTU: typical values are 576 (Internet de facto), 1500 (Ethernet-friendly),
     // 4000 (on some point-to-point links), 4470 (Cisco routers default, FDDI compatible)
-    interfaceEntry->setMtu(par("mtu").longValue());
+    interfaceEntry->setMtu(par("mtu"));
 
     // capabilities
     interfaceEntry->setMulticast(true);
