+++ src/inet/linklayer/lmac/LMacLayer.cc
@@ -146,7 +146,7 @@ InterfaceEntry *LMacLayer::createInterfaceEntry()
     e->setInterfaceToken(address.formInterfaceIdentifier());
 
     // capabilities
-    e->setMtu(par("mtu").longValue());
+    e->setMtu(par("mtu"));
     e->setMulticast(false);
     e->setBroadcast(true);
