+++ src/inet/linklayer/ppp/Ppp.cc
@@ -116,7 +116,7 @@ InterfaceEntry *Ppp::createInterfaceEntry()
 
     // MTU: typical values are 576 (Internet de facto), 1500 (Ethernet-friendly),
     // 4000 (on some point-to-point links), 4470 (Cisco routers default, FDDI compatible)
-    e->setMtu(par("mtu").longValue());
+    e->setMtu(par("mtu"));
 
     // capabilities
     e->setMulticast(true);
