+++ src/inet/networklayer/generic/GenericNetworkProtocol.cc
@@ -456,7 +456,7 @@ void GenericNetworkProtocol::decapsulate(Packet *packet)
 void GenericNetworkProtocol::encapsulate(Packet *transportPacket, const InterfaceEntry *& destIE)
 {
     auto header = makeShared<GenericDatagramHeader>();
-    header->setChunkLength(B(par("headerLength").longValue()));
+    header->setChunkLength(B(par("headerLength")));
     auto l3AddressReq = transportPacket->removeMandatoryTag<L3AddressReq>();
     L3Address src = l3AddressReq->getSrcAddress();
     L3Address dest = l3AddressReq->getDestAddress();
