+++ src/inet/routing/ospfv2/OspfConfigReader.cc
@@ -147,7 +147,7 @@ int OspfConfigReader::getIntAttrOrPar(const cXMLElement& ifConfig, const char *n
     const char *attrStr = ifConfig.getAttribute(name);
     if (attrStr && *attrStr)
         return atoi(attrStr);
-    return par(name).longValue();
+    return par(name);
 }
 
 bool OspfConfigReader::getBoolAttrOrPar(const cXMLElement& ifConfig, const char *name) const
