+++ tests/module/lib/IGMPTester.cc
@@ -125,7 +125,7 @@ void IGMPTester::initialize(int stage)
         MacAddress address("AA:00:00:00:00:01");
         interfaceEntry->setMACAddress(address);
         interfaceEntry->setInterfaceToken(address.formInterfaceIdentifier());
-        interfaceEntry->setMtu(par("mtu").longValue());
+        interfaceEntry->setMtu(par("mtu"));
         interfaceEntry->setMulticast(true);
         interfaceEntry->setBroadcast(true);
